+++ src/data/ipfs.js
@@ -1,3 +1,4 @@
+import { toHHMMSS } from '../utils/time'
 import {
   IPFS_DIRECTORY_MIMETYPE,
   IPFS_DISPLAY_URI_BLACKCIRCLE,
@@ -247,21 +248,4 @@ function getFormatData(item, hash) {
   console.log(data)
 
   return data
-}
-
-function toHHMMSS(sec) {
-  var hours = Math.floor(sec / 3600)
-  var minutes = Math.floor((sec - hours * 3600) / 60)
-  var seconds = sec - hours * 3600 - minutes * 60
-
-  if (hours < 10) {
-    hours = '0' + hours
-  }
-  if (minutes < 10) {
-    minutes = '0' + minutes
-  }
-  if (seconds < 10) {
-    seconds = '0' + seconds
-  }
-  return hours + ':' + minutes + ':' + seconds
 }
