+++ src/utils/compress.js
@@ -1,4 +1,5 @@
 import { createFFmpeg, fetchFile } from '@ffmpeg/ffmpeg'
+import { toHHMMSS } from '../utils/time'
 import { MIMETYPE } from '../constants'
 
 const VIDEO_MIMETYPES = [
@@ -231,21 +232,4 @@ function getVideoMetadata(blob) {
 
     if (DEBUG) document.body.appendChild(video)
   })
-}
-
-function toHHMMSS(sec) {
-  var hours = Math.floor(sec / 3600)
-  var minutes = Math.floor((sec - hours * 3600) / 60)
-  var seconds = sec - hours * 3600 - minutes * 60
-
-  if (hours < 10) {
-    hours = '0' + hours
-  }
-  if (minutes < 10) {
-    minutes = '0' + minutes
-  }
-  if (seconds < 10) {
-    seconds = '0' + seconds
-  }
-  return hours + ':' + minutes + ':' + seconds
 }
